Move UDFContext into an easier to find module
