+++ ibis/bigquery/udf/core.py
@@ -13,26 +13,8 @@ import six
 
 import ibis.expr.datatypes as dt
 
-from ibis.bigquery.datatypes import (
-    ibis_type_to_bigquery_type, TypeTranslationContext
-)
 from ibis.bigquery.udf.find import find_names
 from ibis.bigquery.udf.rewrite import rewrite
-
-
-class UDFContext(TypeTranslationContext):
-    __slots__ = ()
-
-
-@ibis_type_to_bigquery_type.register(dt.Integer, UDFContext)
-def trans_integer(t, context):
-    # JavaScript does not have integers, only a Number class. BigQuery doesn't
-    # behave as expected with INT64 inputs or outputs
-    raise TypeError(
-        'BigQuery does not support INT64 as an input or output type for UDFs. '
-        'Replace INT64 with FLOAT64 in your UDF signature specification and '
-        'cast any inputs to FLOAT64.'
-    )
 
 
 class SymbolTable(ChainMap):
